+++ src/main/java/com/woowacamp/soolsool/core/cart/domain/CartItem.java
@@ -1,12 +1,11 @@
 package com.woowacamp.soolsool.core.cart.domain;
 
-import static com.woowacamp.soolsool.global.exception.GlobalErrorCode.NO_CONTENT;
+import static com.woowacamp.soolsool.global.code.GlobalErrorCode.NO_CONTENT;
 import static lombok.AccessLevel.PROTECTED;
 
 import com.woowacamp.soolsool.core.cart.domain.converter.CartItemQuantityConverter;
 import com.woowacamp.soolsool.core.cart.domain.vo.CartItemQuantity;
 import com.woowacamp.soolsool.core.liquor.domain.Liquor;
-import com.woowacamp.soolsool.global.code.GlobalErrorCode;
 import com.woowacamp.soolsool.global.common.BaseEntity;
 import com.woowacamp.soolsool.global.exception.SoolSoolException;
 import java.math.BigInteger;
