+++ src/main/java/com/woowacamp/soolsool/core/liquor/domain/Liquor.java
@@ -1,5 +1,7 @@
 package com.woowacamp.soolsool.core.liquor.domain;
 
+import static com.woowacamp.soolsool.global.code.GlobalErrorCode.NO_CONTENT;
+
 import com.woowacamp.soolsool.core.liquor.domain.converter.LiquorAlcoholConverter;
 import com.woowacamp.soolsool.core.liquor.domain.converter.LiquorBrandConverter;
 import com.woowacamp.soolsool.core.liquor.domain.converter.LiquorImageUrlConverter;
@@ -17,7 +19,6 @@ import com.woowacamp.soolsool.core.liquor.domain.vo.LiquorStock;
 import com.woowacamp.soolsool.core.liquor.domain.vo.LiquorVolume;
 import com.woowacamp.soolsool.core.liquor.dto.LiquorModifyRequest;
 import com.woowacamp.soolsool.global.common.BaseEntity;
-import com.woowacamp.soolsool.global.exception.GlobalErrorCode;
 import com.woowacamp.soolsool.global.exception.SoolSoolException;
 import java.math.BigInteger;
 import java.time.LocalDateTime;
@@ -151,7 +152,7 @@ public class Liquor extends BaseEntity {
 
     private void validateIsNotNullableCategory(final Object... objects) {
         if (Arrays.stream(objects).anyMatch(Objects::isNull)) {
-            throw new SoolSoolException(GlobalErrorCode.NO_CONTENT);
+            throw new SoolSoolException(NO_CONTENT);
         }
     }
