+++ src/main/java/com/woowacamp/soolsool/core/receipt/config/ReceiptConfig.java
@@ -1,4 +1,4 @@
-package com.woowacamp.soolsool.core.liquor.config;
+package com.woowacamp.soolsool.core.receipt.config;
 
 import com.woowacamp.soolsool.core.receipt.ReceiptStatusTypeConverter;
 import org.springframework.context.annotation.Configuration;
@@ -6,13 +6,10 @@ import org.springframework.format.FormatterRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
-public class LiquorConfig implements WebMvcConfigurer {
+public class ReceiptConfig implements WebMvcConfigurer {
 
     @Override
     public void addFormatters(final FormatterRegistry registry) {
-        registry.addConverter(new LiquorBrewTypeConverter());
-        registry.addConverter(new LiquorRegionTypeConverter());
-        registry.addConverter(new LiquorStatusTypeConverter());
         registry.addConverter(new ReceiptStatusTypeConverter());
     }
 }
