+++ src/test/java/com/woowacamp/soolsool/core/order/domain/vo/OrderPriceTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
-import com.woowacamp.soolsool.global.exception.ShoppingException;
+import com.woowacamp.soolsool.global.exception.SoolSoolException;
 import java.math.BigInteger;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -27,26 +27,25 @@ class OrderPriceTest {
 
     @ParameterizedTest
     @ValueSource(strings = {"-1", "0"})
-    @DisplayName("주문 가격이 0 이하일 경우 ShoppingException을 던진다.")
+    @DisplayName("주문 가격이 0 이하일 경우 SoolSoolException을 던진다.")
     void createFailWithInvalidPrice(String invalidPrice) {
         // given
         BigInteger price = new BigInteger(invalidPrice);
 
         // when & then
         assertThatThrownBy(() -> new OrderPrice(price))
-            .isExactlyInstanceOf(ShoppingException.class)
+            .isExactlyInstanceOf(SoolSoolException.class)
             .hasMessage("주문 가격은 0 이하일 수 없습니다.");
     }
 
     @Test
-    @DisplayName("주문 가격이 null일 경우 ShoppingException을 던진다.")
+    @DisplayName("주문 가격이 null일 경우 SoolSoolException을 던진다.")
     void createFailWithNull() {
         // given
 
-
         // when & then
         assertThatThrownBy(() -> new OrderPrice(null))
-            .isExactlyInstanceOf(ShoppingException.class)
+            .isExactlyInstanceOf(SoolSoolException.class)
             .hasMessage("주문 가격은 null일 수 없습니다.");
     }
